worker/requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9

